Ensure recv_ready() always returns False for sftp
We should allow the SFTP subsystem handler to handle the
packets, which is a threaded server that runs elsewhere.

Otherwise, we have two read(2) on the same socket, causing
a sort of ping-pong effect of who handles and who mishandles
the next request.

We also modify 'kind' variable when sftp is requested, to
'sftp', this also ensures 'matrix_sftp' is called when specified
(and very well should be!)